requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
pyyaml==6.0.2
lxml==5.3.0
beautifulsoup4==4.12.3
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
